STAC-12961: Move health api to health util
