+++ stackstate_checks_base/stackstate_checks/base/__init__.py
@@ -11,7 +11,7 @@ from .utils.common import ensure_string, ensure_unicode, to_string
 from .utils.identifiers import Identifiers
 from .utils.telemetry import MetricStream, MetricHealthChecks, EventStream, EventHealthChecks, HealthState,\
     ServiceCheckStream, ServiceCheckHealthChecks, TopologyEventContext, SourceLink, Event
-from .utils.health import Health, HealthStream, HealthStreamUrn, HealthType
+from .utils.health_api import Health, HealthStream, HealthStreamUrn, HealthType
 from .utils.agent_integration_test_util import AgentIntegrationTestUtil
 from .utils.persistent_state import StateDescriptor, StateManager, StateNotPersistedException, \
     StateCorruptedException, StateReadException
